feat: share admin proof state
move read-only admin proof state and D1 proof readers into @anipotts/content/admin so apps/admin only renders the proof route. this keeps content and passkey proof read-only while putting admin state contracts in the shared package.

verify: pnpm --filter @anipotts/content build

verify: pnpm turbo typecheck --filter=@anipotts/admin...

verify: pnpm turbo build --filter=@anipotts/admin...

verify: pnpm test:content-platform

verify: pnpm test:admin-routes

verify: pnpm test:deploy-targets

verify: pnpm test:security-review

verify: pnpm --silent proof:admin-content

verify: pnpm --silent proof:admin-passkey

verify: pnpm format:check

verify: git diff --check

verify: node scripts/ci/security-review.mjs --required /tmp/anipotts-com-changed-files.txt
