+++ packages/content/src/admin/index.ts
@@ -1,4 +1,5 @@
 export * from "./content.js";
 export * from "./newsletter.js";
 export * from "./operations.js";
+export * from "./proof.js";
 export * from "./source-content.js";
