+++ scripts/ci/content-platform.test.mjs
@@ -3,8 +3,11 @@
 import assert from "node:assert/strict";
 import { execFileSync } from "node:child_process";
 import {
+  countProofEntries,
   contentInventorySource,
+  proofSource,
   recordsFromSourceModules,
+  readProofEntries,
   summarizeSourceContentRecords,
 } from "../../packages/content/dist/admin/index.js";
 import {
@@ -120,6 +123,30 @@ assert.equal(
   ),
   "read_only_static_plus_d1_page_content",
   "apps/admin must be able to import @anipotts/content/admin from the built package export",
+);
+
+assert.equal(proofSource.mode, "read_only_d1_plus_runtime_metadata");
+assert.equal(proofSource.live_writes, "disabled");
+
+const proofEntriesWithoutDb = await readProofEntries(undefined);
+assert.deepEqual(
+  countProofEntries(proofEntriesWithoutDb),
+  {
+    total: 6,
+    verified: 4,
+    blocked: 1,
+    pending: 1,
+  },
+  "proof exports must preserve read-only fallback status without an app D1 binding",
+);
+assert.ok(
+  proofEntriesWithoutDb.some(
+    (entry) =>
+      entry.id === "proof.admin.passkey-enrollment" &&
+      entry.status === "blocked" &&
+      entry.next_safe_action.includes("DB binding"),
+  ),
+  "proof fallback must keep Access removal blocked when passkey proof is unavailable",
 );
 
 assert.deepEqual(
